requirements.txt
pymupdf>=1.27.0,<2
pydantic>=2.0,<3
numpy>=1.26.0,<3
openai>=1.40.0,<2
cohere>=5.0.0,<7
sentence-transformers>=3.0.0,<5
fastapi>=0.115.0,<1
uvicorn[standard]>=0.30.0,<1
httpx>=0.27.0,<1
slowapi>=0.1.9,<1
python-multipart>=0.0.9,<1
redis>=5.0.0,<6
rq>=2.0.0,<3
psycopg[binary]>=3.1.0,<4
pgvector>=0.3.0,<1
boto3>=1.35.0,<2
python-dotenv>=1.0.0,<2
